feat: Install TS, convert files, add scripts and test out build
