+++ server/src/singletons/WebSocketManager.ts
@@ -2,28 +2,30 @@
 import { WebSocket } from 'ws';
 
 class WebSocketManager {
+  connections: Set<WebSocket>;
+
   constructor() {
-    this.connections = new Set();
+    this.connections = new Set<WebSocket>();
   }
 
   get getConnections() {
     return this.connections;
   }
 
-  addConnection(newConnection) {
+  addConnection(newConnection: WebSocket) {
     if (newConnection instanceof WebSocket) {
       this.connections.add(newConnection);
     }
   }
 
-  removeConnection(connection) {
+  removeConnection(connection: WebSocket) {
     if (connection instanceof WebSocket) {
       this.connections.delete(connection);
       console.log('Connection removed');
     }
   }
 
-  pingConnection(connection, message) {
+  pingConnection(connection: WebSocket, message: string) {
     if (connection instanceof WebSocket) {
       const stringifiedMessage = JSON.stringify({
         pingMsg: message,
@@ -32,15 +34,15 @@ class WebSocketManager {
     }
   }
 
-  broadcast(payload) {
+  broadcast(payload: string) {
     for (const connection of this.connections) {
-      if (connection instanceof WebSocket && connection.readyState === WebSocket.OPEN) {
+      if (connection.readyState === WebSocket.OPEN) {
         connection.send(payload);
       }
     }
   }
 
-  receiveDataFromConnection(data, callback) {
+  receiveDataFromConnection(data: string, callback: (parsedData: any) => void) {
     const parsedData = JSON.parse(data);
     callback(parsedData);
   }
