+++ server/src/types/RetroNoteData.ts
@@ -1,0 +1,5 @@
+export type RetroNoteData = {
+  creatorId: string;
+  text: string;
+  topic: string;
+};
